+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.1
 
+- Added Menu button to bottom right
 - Fixed results view not showing after quiz completed
 
 ## 0.0.1