+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.1
 
+- Removed "Show Results" on last question in list because it would set turn to completed
 - Added `export_quiz "<QUIZ_NAME>"` command
 - Updated to PrismJS `1.29.0`
 - Refactored problem submission to apply everywhere